massive commit. more info in description new timebase, changes to song parsing, start font size changes, new smashers, disable FXAA, new logger, and idk what else i did :sob:

+++ Encore/src/users/player.h
@@ -408,7 +408,7 @@ public:
     bool LeftyFlip;
     bool Online;
     int ActiveSlot;
-    int HighwayLength;
+    float HighwayLength;
 
     void ResetGameplayStats();
